refactor: ConfigクラスとConfigBuilderの型をunsigned intからsize_tに変更
- Configクラスのメンバ変数とgetterメソッドの型を変更
- ConfigBuilderのsetterメソッドのパラメータ型を変更
- より適切な型表現により、コードの一貫性を向上

+++ src/Config/ConfigBuilder.cpp
@@ -81,23 +81,23 @@ ConfigBuilder::~ConfigBuilder() {}
 Config ConfigBuilder::build() const {
 	return Config(_listens, _locations, _accessLogs, _errorLogs,
 				  _maxRequestBodySize, _hasBiggestRequestBodySize,
-				  _biggestRequestBodySize, _timeoutSec,
-				  _maxEvents, _requestHeaderTimeoutSec, _requestBodyTimeoutSec,
+				  _biggestRequestBodySize, _timeoutSec, _maxEvents,
+				  _requestHeaderTimeoutSec, _requestBodyTimeoutSec,
 				  _errorPages);
 }
 
-void ConfigBuilder::setMaxRequestBodySize(const unsigned int size) {
+void ConfigBuilder::setMaxRequestBodySize(const size_t size) {
 	_maxRequestBodySize = size;
 }
 
-void ConfigBuilder::setBiggestRequestBodySize(bool hasValue, unsigned int size) {
+void ConfigBuilder::setBiggestRequestBodySize(bool hasValue, size_t size) {
 	_hasBiggestRequestBodySize = hasValue;
 	_biggestRequestBodySize = size;
 }
 
-void ConfigBuilder::setTimeoutSec(const unsigned int sec) { _timeoutSec = sec; }
+void ConfigBuilder::setTimeoutSec(const size_t sec) { _timeoutSec = sec; }
 
-void ConfigBuilder::setMaxEvents(const unsigned int maxEvents) {
+void ConfigBuilder::setMaxEvents(const size_t maxEvents) {
 	_maxEvents = maxEvents;
 }
 
@@ -209,10 +209,10 @@ void ConfigBuilder::setServerDefaultSession(bool enable) {
 	_locations[_defaultLocationKey].session = enable;
 }
 
-void ConfigBuilder::setRequestHeaderTimeoutSec(const unsigned int sec) {
+void ConfigBuilder::setRequestHeaderTimeoutSec(const size_t sec) {
 	_requestHeaderTimeoutSec = sec;
 }
 
-void ConfigBuilder::setRequestBodyTimeoutSec(const unsigned int sec) {
+void ConfigBuilder::setRequestBodyTimeoutSec(const size_t sec) {
 	_requestBodyTimeoutSec = sec;
 }
